test: consolidate local runtime contracts
